# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "REST API for BloodConnect donor, blood request, donation, and hospital services.",
        "title": "BloodConnect Donor Service API",
        "contact": {},
        "version": "1.0"
    },
    "host": "localhost:1324",
    "basePath": "/api/v1",
    "paths": {
        "/blood-requests": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Get blood requests owned by the authenticated requester",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Blood Requests"
                ],
                "summary": "Get blood requests",
                "responses": {
                    "200": {
                        "description": "Success",
                        "schema": {
                            "$ref": "#/definitions/helper.BloodRequestListSwaggoResponse"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/helper.ErrorSwaggoResponse"
                        }
                    },
                    "403": {
                        "description": "Forbidden",
                        "schema": {
                            "$ref": "#/definitions/helper.ErrorSwaggoResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/helper.ErrorSwaggoResponse"
                        }
                    }
                }
            },
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Blood Requests"
                ],
                "parameters": [
                    {
                        "description": "Blood request data",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/domain.BloodRequestReq"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Blood request created",
                        "schema": {
                            "$ref": "#/definitions/helper.BloodRequestSwaggoResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/helper.ErrorSwaggoResponse"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/helper.ErrorSwaggoResponse"
                        }
                    },
                    "403": {
                        "description": "Forbidden",
                        "schema": {
                            "$ref": "#/definitions/helper.ErrorSwaggoResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/helper.ErrorSwaggoResponse"
                        }
                    }
                }
            }
        },
        "/blood-requests/{id}": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Get a blood request by ID",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Blood Requests"
                ],
                "summary": "Get blood request detail",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Blood Request ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Success",
                        "schema": {
                            "$ref": "#/definitions/helper.BloodRequestSwaggoResponse"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/helper.ErrorSwaggoResponse"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/helper.ErrorSwaggoResponse"
                        }
                    },
                    "422": {
                        "description": "Unprocessable Content",
                        "schema": {
                            "$ref": "#/definitions/helper.ErrorSwaggoResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/helper.ErrorSwaggoResponse"
                        }
                    }
                }
            }
        },
        "/blood-requests/{id}/cancel": {
            "put": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Cancel a blood request owned by the authenticated requester",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Blood Requests"
                ],
                "summary": "Cancel blood request",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Blood Request ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Blood request cancelled",
                        "schema": {
                            "$ref": "#/definitions/helper.BloodRequestSwaggoResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/helper.ErrorSwaggoResponse"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/helper.ErrorSwaggoResponse"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/helper.ErrorSwaggoResponse"
                        }
                    },
                    "422": {
                        "description": "Unprocessable Content",
                        "schema": {
                            "$ref": "#/definitions/helper.ErrorSwaggoResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/helper.ErrorSwaggoResponse"
                        }
                    }
                }
            }
        },
...
        "/blood-requests/{id}/matches": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Get donor matches for a blood request",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Blood Requests"
                ],
                "summary": "Get donor matches",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Blood Request ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Success",
                        "schema": {
                            "$ref": "#/definitions/helper.DonorMatchListSwaggoResponse"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/helper.ErrorSwaggoResponse"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/helper.ErrorSwaggoResponse"
                        }
                    },
                    "422": {
                        "description": "Unprocessable Content",
                        "schema": {
                            "$ref": "#/definitions/helper.ErrorSwaggoResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/helper.ErrorSwaggoResponse"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
        "domain.BloodRequestReq": {
            "type": "object",
            "properties": {
                "blood_type": {
                    "type": "string"
                },
                "city": {
                    "type": "string"
                },
                "hospital_external_id": {
                    "type": "string"
                },
                "hospital_name": {
                    "type": "string"
                },
                "latitude": {
                    "type": "number"
                },
                "longitude": {
                    "type": "number"
                },
                "needed_at": {
                    "type": "string"
                },
                "notes": {
                    "type": "string"
                },
                "quantity": {
                    "type": "integer"
                },
                "urgency": {
                    "type": "string"
                }
            }
        },
...
        "entity.BloodRequest": {
            "type": "object",
            "properties": {
                "blood_type": {
                    "type": "string"
                },
                "city": {
                    "type": "string"
                },
                "created_at": {
                    "type": "string"
                },
                "donor_matches": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/entity.DonorMatch"
                    }
                },
                "hospital_external_id": {
                    "type": "string"
                },
                "hospital_name": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "latitude": {
                    "type": "number"
                },
                "longitude": {
                    "type": "number"
                },
                "needed_at": {
                    "type": "string"
                },
                "notes": {
                    "type": "string"
                },
                "quantity": {
                    "type": "integer"
                },
                "requester_id": {
                    "type": "string"
                },
                "status": {
                    "type": "string"
                },
                "updated_at": {
                    "type": "string"
                },
                "urgency": {
                    "type": "string"
                }
            }
        },
...
        "entity.DonorMatch": {
            "type": "object",
            "properties": {
                "blood_request": {
                    "$ref": "#/definitions/entity.BloodRequest"
                },
                "blood_request_id": {
                    "type": "string"
                },
                "created_at": {
                    "type": "string"
                },
                "distance_km": {
                    "type": "number"
                },
                "donor_id": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "status": {
                    "type": "string"
                },
                "updated_at": {
                    "type": "string"
                }
            }
        },
        "helper.BloodRequestListSwaggoResponse": {
            "type": "object",
            "properties": {
                "data": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/entity.BloodRequest"
                    }
                },
                "message": {
                    "type": "string"
                },
                "status": {
                    "type": "string"
                }
            }
        },
        "helper.BloodRequestSwaggoResponse": {
            "type": "object",
            "properties": {
                "data": {
                    "$ref": "#/definitions/entity.BloodRequest"
                },
                "message": {
                    "type": "string"
                },
                "status": {
                    "type": "string"
                }
            }
        },
...
        "helper.DonorMatchListSwaggoResponse": {
            "type": "object",
            "properties": {
                "data": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/entity.DonorMatch"
                    }
                },
                "message": {
                    "type": "string"
                },
                "status": {
                    "type": "string"
                }
            }
        },
...
        "helper.ErrorDetailSwaggoResponse": {
            "type": "object",
            "properties": {
                "detail": {
                    "type": "string"
                },
                "error_code": {
                    "type": "string"
                },
                "message": {
                    "type": "string"
                }
            }
        },
        "helper.ErrorSwaggoResponse": {
            "type": "object",
            "properties": {
                "error": {
                    "$ref": "#/definitions/helper.ErrorDetailSwaggoResponse"
                },
                "status": {
                    "type": "string"
                }
            }
        },
...
    },
    "securityDefinitions": {
        "BearerAuth": {
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}